backend/requirements.txt
fastapi
uvicorn
google-genai
edge-tts
aiohttp
python-multipart
python-dotenv
